Force images to disk to avoid corruption
Sometimes images downloaded from glance that are then converted to
raw format end up corrupt.  This is becasue qemu-img uses fiemap to
read blocks from the source image but is not using FIEMAP_FLAG_SYNC,
so blocks and extents that are in the kernel page_cache but not on
disk are skipped.

This results in corrupt images.

Create a new 'workarounds' group and a flag (qemu_force_sync) to
force fsync() in GlanceImageService.download().  This flag can also
be used in other code paths when we know that they are subject to the
same fiemap issue.

'qemu_force_sync' defaults to True.  It is intended that
distributions that know they have a fixed qemu can add the
appropriate config options to avoid the
fsync() cost.

As mentioned in the code distro status can be tracked at:
    https://wiki.openstack.org/wiki/Bug1368815

The impact on the other tests in TestDownloadNoDirectUri is
regretable but AFAICT unavoidable.

Change-Id: Ieb33489808a71c1fb40cd6835f0bb753dd69e293
Closes-Bug: #1368815
Co-Authored-By: Davanum Srinivas <dims@linux.vnet.ibm.com>